+++ R/ms_site.R
@@ -17,6 +17,7 @@
 #' - `list_drives(filter=NULL, n=Inf)`: List the drives (shared document libraries) associated with this site.
 #' - `get_drive(drive_name, drive_id)`: Retrieve a shared document library for this site. If the name and ID are not specified, this returns the default document library.
 #' - `list_subsites(filter=NULL, n=Inf)`: List the subsites of this site.
+#' - `get_pages(type = c("sitePage", "page"))`: Returns SharePoint pages for this site.
 #' - `get_lists(filter=NULL, n=Inf)`: Returns the lists that are part of this site.
 #' - `get_list(list_name, list_id)`: Returns a specific list, either by name or ID.
 #' - `get_group()`: Retrieve the Microsoft 365 group associated with the site, if it exists. A site that backs a private Teams channel will not have a group associated with it.
@@ -44,6 +45,7 @@
 #' }
 #' @format An R6 object of class `ms_site`, inheriting from `ms_object`.
 #' @export
+#' @importFrom curl curl_escape
 ms_site <- R6::R6Class("ms_site", inherit=ms_object,
 
 public=list(
@@ -107,6 +109,43 @@ public=list(
         if(length(res) != 1)
             stop("Unable to get Microsoft 365 group", call.=FALSE)
         az_group$new(self$token, self$tenant, res[[1]])
+    },
+
+    get_analytics=function()
+    {
+      self$do_operation("analytics")
+    },
+
+    get_pages=function(type=c("sitePage", "page"))
+    {
+      type <- match.arg(type)
+
+      op <- switch(type,
+        page = "pages",
+        sitePage = "pages/microsoft.graph.sitePage"
+      )
+
+      self$do_operation(op)
+    },
+
+    # https://learn.microsoft.com/en-us/graph/api/site-list-columns
+    # FIXME: Add support for OData parameters
+    list_columns=function()
+    {
+      self$do_operation("columns")
+    },
+
+    # https://learn.microsoft.com/en-us/graph/api/site-list-permissions?view=graph-rest-beta&tabs=http
+    # FIXME: Add support for OData parameters
+    list_permissions=function()
+    {
+      self$do_operation("permissions")
+    },
+
+    # https://learn.microsoft.com/en-us/graph/api/site-list-contenttypes?view=graph-rest-beta&tabs=http
+    list_content_types=function()
+    {
+      self$do_operation("contentTypes")
     },
 
     print=function(...)